Extracting `FieldActions` component.
